+++ frontend/spotted-apple/features/create-account-form-feature/form.tsx
@@ -1,8 +1,8 @@
 'use client';
 
 import { ChangeEvent, useState, } from 'react'
-import { FormField, FormFieldError, SubmitButton } from '@/app/lib/forms/components'
-import { ButtonDisplay, InputFields, InputFieldErrorState} from '@/app/lib/forms/types'
+import { FormField, FormFieldError, SubmitButton } from '@/lib/forms/components'
+import { ButtonDisplay, InputFields, InputFieldErrorState} from '@/lib/forms/types'
 import { CreateAccountFormProps } from '@/features/create-account-form-feature/types';
 
 export default function CreateAccountForm({state, formAction}: CreateAccountFormProps) {
